+++ Model/Connect.php
@@ -646,7 +646,7 @@ class Connect extends \Magento\Payment\Model\Method\AbstractMethod
                 "plugin" => [
                     "shop" => $magentoInfo->getName() . ' ' . $magentoInfo->getVersion() . ' ' . $magentoInfo->getEdition(),
                     "shop_version" => $magentoInfo->getVersion(),
-                    "plugin_version" => ' - Plugin 1.13.0',
+                    "plugin_version" => ' - Plugin 1.13.1',
                     "partner" => "MultiSafepay",
                 ],
                 "gateway_info" => [